smarter statistics analysis in recognizer
