+++ src/main/python/Card.py
@@ -4,24 +4,15 @@ class Card():
     def __init__(self, profile, exists):
         self.exists = exists
         self.profile = profile
-        self.number= None
-        self.color= None
-        self.compCenter= None
-        self.compPoints= None
-        self.compContour= None
+        self.number = None
+        self.color = None
+        self.compCenter = None
+        self.compPoints = None
+        self.compContour = None
         self.counter = 1
 
-        #Is set before comparison when validating which card the filmed card represents
-        self.rankImg = []   # 125*70
-        self.suitImg = []   # 100*70
-        
-        self.rankAndSuitContourImgs = [] # 125*70
-
-
-        self.best_rank_match = None #name of the best matched rank
-        self.best_suit_match = None #name of the best matched suit
-        self.rank_match_difference_value = 0 #difference between the best matching rank image and the checking image
-        self.suit_match_difference_value = 0 #difference between the best matching rank image and the checking image
+        self.suit = "[no suit set]"
+        self.rank = "[no rank set]"
 
     # Author : Hans
     def initiateCard(self, contour, points):
@@ -32,6 +23,7 @@ class Card():
     # Author : Hans
     def setProfile(self, image):
         self.profile = image
+
 
 """class Card():
